Create helper functions for status calculation
